questions_app/requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.21.1
requests==2.26.0
